requirements.txt
watchdog>=2.1.9
pypdf>=3.7.0
click>=8.0.0
swarm-python>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0  # For embeddings
requests>=2.31.0  # For Ollama API
